+++ src/app.js
@@ -3,8 +3,9 @@ import dotenv from "dotenv";
 import TestRestController from "./testModule/interface/TestRestController";
 import MockTestRepository from "./testModule/infrastructure/MockTestRepository";
 import TestService from "./testModule/core/TestService";
-import auth from "./userModule/routes/auth";
-import hash from "./userModule/routes/hash";
+import auth from "../src/userModule/interface/auth";
+import changePas from "../src/userModule/interface/changePas";
+import newPas from "../src/userModule/interface/newPas";
 
 dotenv.config();
 const app = express();
@@ -19,7 +20,8 @@ const testApi = new TestRestController(testServiceImplementation);
 //wiring up the routes
 app.use("/tests",testApi.router);
 app.use('/auth', auth);
-app.use('/hash', hash);
+app.use('/changePas', changePas);
+app.use('/newPas', newPas);
 
 app.listen(PORT, () => {
 	console.log(`Listening on port ${PORT}...`);
